move read_cookies_from_file back to generate_cookies.py
