+++ app/modules/instagram_content_loader.py
@@ -1,7 +1,6 @@
 import requests
 from config.tgram_bot_logger import write_log
-from .generate_cookies import generate_cookies
-from .tgram_bot_helper import read_cookies_from_file
+from .generate_cookies import generate_cookies, read_cookies_from_file
 
 retry_count = 0
